Update to upstream version 3.2
6fcb58f40dae8228074591e3262d5db97db3e2d7

+++ editor/editor_properties.h
@@ -278,14 +278,14 @@ class EditorPropertyInteger : public EditorProperty {
 
     EditorSpinSlider *spin;
     bool setting;
-    void _value_changed(double p_val);
+    void _value_changed(int64_t p_val);
 
 protected:
     static void _bind_methods();
 
 public:
     void update_property() override;
-    void setup(int p_min, int p_max, int p_step, bool p_allow_greater, bool p_allow_lesser);
+    void setup(int64_t p_min, int64_t p_max, int64_t p_step, bool p_allow_greater, bool p_allow_lesser);
     EditorPropertyInteger();
 };
